[add] base FrameDataset과 raw bgr image format read test code (#4)
