--- src/domain/__init__.py
@@ -1,11 +1,0 @@
-from .entity import ExtractedFaceImage, FaceMask, Frame, Video
-from .extraction import get_frames_from_video, make_frame_list
-
-__all__ = [
-    "Video",
-    "Frame",
-    "ExtractedFaceImage",
-    "FaceMask",
-    "get_frames_from_video",
-    "make_frame_list",
-]
